+++ R/Corpus_class.R
@@ -3,16 +3,53 @@ setOldClass("Corpus")
 
 #' Corpus class.
 #' 
-#' @field corpus character vector, CWB corpus
-#' @field encoding encoding of the corpus
-#' @field count data.table with counts
+#' The R6 \code{Corpus} class offers a set of methods to retrieve and manage CWB
+#' indexed corpora.
+#' 
+#' @field corpus character vector (length 1), a CWB corpus
+#' @field encoding encoding of the corpus (typically 'UTF-8' or 'latin1'), assigned
+#' automatically upon initialization of the corpus
+#' @field cpos a two-column \code{matrix} with regions of a corpus underlying the
+#' s-attributes of the \code{data.table} in field \code{sAttributes}
+#' @field sAttributes a \code{data.table} with the values of a set of sAttributes
+#' @field stat a \code{data.table} with counts
+#' 
+#' @section Arguments:
+#' \describe{
+#'   \item{corpus}{a corpus}
+#'   \item{pAttribute}{p-attribute, to perform count}
+#'   \item{sAttributes}{s-attributes}
+#'   \item{id2str}{logical, whether to turn token ids into strings upon counting}
+#'   \item{as.html}{logical}
+#' }
+#' 
 #' @section Methods:
 #' \describe{
+#'   \item{\code{initialize(corpus, pAttribute = NULL, sAttributes = NULL)}}{Initialize a new object of class \code{Corpus}.}
 #'   \item{\code{count(pAttribute = getOption("polmineR.pAttribute"), id2str = TRUE)}}{Perform counts.}
+#'   \item{\code{as.partition()}}{turn \code{Corpus} into a partition}
+#'   \item{\code{getInfo(as.html = FALSE)}}{}
+#'   \item{\code{showInfo()}}{}
 #' }
 #' 
 #' @rdname Corpus-class
 #' @export Corpus
+#' @examples
+#' use("polmineR")
+#' REUTERS <- Corpus$new("REUTERS")
+#' infofile <- REUTERS$getInfo()
+#' if (interactive()) REUTERS$showInfo()
+#' 
+#' # use Corpus class to manage counts
+#' REUTERS <- Corpus$new("REUTERS", pAttribute = "word")
+#' REUTERS$stat
+#' 
+#' # use Corpus class for creating partitions
+#' REUTERS <- Corpus$new("REUTERS", sAttributes = c("id", "places"))
+#' usa <- partition(REUTERS, places = "usa")
+#' sa <- partition(REUTERS, places = "saudi-arabia", regex = TRUE)
+#' 
+#' reut <- REUTERS$as.partition()
 Corpus <- R6Class(
   
   "Corpus",
@@ -21,17 +58,66 @@ Corpus <- R6Class(
     
     corpus = NULL,
     encoding = NULL,
+    cpos = NULL,
     pAttribute = NULL,
+    sAttributes = NULL,
     size = NULL,
     stat = data.table(),
     
-    initialize = function(corpus, pAttribute = NULL){
+    initialize = function(corpus, pAttribute = NULL, sAttributes = NULL){
+      stopifnot(is.character(corpus), length(corpus) == 1)
       if (!corpus %in% polmineR::corpus()[["corpus"]]) warning("corpus may not be available")
       self$corpus <- corpus
       self$encoding <- getEncoding(corpus)
       self$size <- size(corpus)
-      if (!is.null(pAttribute)) self$count(pAttribute)
-    }, 
+      if (!is.null(pAttribute)){
+        stopifnot(pAttribute %in% pAttributes(corpus))
+        self$count(pAttribute)
+      }
+      if (!is.null(sAttributes)){
+        dt <- decode(corpus, sAttribute = sAttributes)
+        self$cpos <- as.matrix(dt[, 1:2])
+        self$sAttributes <- dt[, 3:ncol(dt)]
+        self$sAttributes[["struc"]] <- 0:(nrow(self$sAttributes) - 1)
+        setcolorder(self$sAttributes, neworder = c(ncol(self$sAttributes), 1:(ncol(self$sAttributes) - 1)))
+      }
+    },
+    
+    # summary = function(sAttributes = c(period = "text_lp", date = "text_date", dummy = "text_id")){
+    #   # generate bundle, each of which will be evaluated in consecutive steps
+    #   lpCluster <- partitionBundle(
+    #     self$corpus,
+    #     def = setNames(list(".*"), sAttributes["dummy"]),
+    #     var = setNames(list(NULL), sAttributes["period"]),
+    #     regex = TRUE, mc = FALSE
+    #   )
+    #   # extract data
+    #   corpusData <- lapply(
+    #     lpCluster@objects,
+    #     function(x) {
+    #       dates <- as.Date(sAttributes(x, sAttributes["date"]))
+    #       c(
+    #         first = as.character(min(dates, na.rm = TRUE)),
+    #         last = as.character(max(dates, na.rm = TRUE)),
+    #         no_token = size(x)
+    #       )
+    #     })
+    #   
+    #   # assemling data.frame
+    #   tab <- data.frame(do.call(rbind, corpusData))
+    #   tab[,"no_token"] <- as.integer(as.vector(tab[,"no_token"]))
+    #   tab <- cbind(tab, avg_no_token = round(tab[,"no_token"]/tab[,"no"], digits = 0))
+    #   
+    #   tab_all <- data.frame(
+    #     first = tab[1, "first"],
+    #     last = tab[nrow(tab), "last"],
+    #     no_token = sum(tab[, "no_token"]),
+    #     avg_no_token = round(sum(tab[,"no_token"])/sum(tab[, "no"]), digits = 0),
+    #     row.names = "ALL"
+    #   )
+    #   tab <- rbind(tab, tab_all)
+    #   tab
+    # },
     
     count = function(pAttribute = getOption("polmineR.pAttribute"), id2str = TRUE){
       self$pAttribute <- pAttribute
@@ -55,7 +141,7 @@ Corpus <- R6Class(
         cpos = matrix(c(0, (size(self$corpus) - 1)), nrow = 1),
         stat = self$stat,
         size = self$size,
-        pAttribute = self$pAttribute
+        pAttribute = if (is.null(self$pAttribute)) character() else self$pAttribute
       )
     },
     
@@ -81,4 +167,3 @@ Corpus <- R6Class(
     }
   )
 )
-